--- a/ast/api_compat.go
+++ b/ast/api_compat.go
@@ -1,4 +1,4 @@
-// +build !amd64 go1.21
+// +build !amd64 !go1.16 go1.21
 
 /*
  * Copyright 2022 ByteDance Inc.