--- a/decoder/decoder_compat.go
+++ b/decoder/decoder_compat.go
@@ -1,3 +1,4 @@
+//go:build !amd64 || !go1.16 || go1.21
 // +build !amd64 !go1.16 go1.21
 
 /*
@@ -14,17 +15,19 @@
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
-*/
+ */
 
 package decoder
 
 import (
-     `encoding/json`
-     `bytes`
-     `reflect`
-     `github.com/bytedance/sonic/internal/native/types`
-     `github.com/bytedance/sonic/option`
-     `io`
+    `bytes`
+    `encoding/json`
+    `io`
+    `reflect`
+    `unsafe`
+
+    `github.com/bytedance/sonic/internal/native/types`
+    `github.com/bytedance/sonic/option`
 )
 
 const (
@@ -194,3 +197,17 @@ func (self *StreamDecoder) Decode(val interface{}) (err error) {
    return dec.Decode(val)
 }
 
+// SyntaxError represents json syntax error
+type SyntaxError json.SyntaxError
+
+// Description
+func (s SyntaxError) Description() string {
+     return (*json.SyntaxError)(unsafe.Pointer(&s)).Error()
+}
+// Error
+func (s SyntaxError) Error() string {
+     return (*json.SyntaxError)(unsafe.Pointer(&s)).Error()
+}
+
+// MismatchTypeError represents dismatching between json and object
+type MismatchTypeError json.UnmarshalTypeError