--- a/encoder/encoder_compat.go
+++ b/encoder/encoder_compat.go
@@ -27,6 +27,10 @@ import (
     `github.com/bytedance/sonic/option`
 )
 
+func init() {
+    println("WARNING: sonic only supports Go1.16~1.20 && CPU amd64, but your environment is not suitable")
+}
+
 // Options is a set of encoding options.
 type Options uint64