--- a/fuzz/Makefile
+++ b/fuzz/Makefile
@@ -5,10 +5,11 @@ fuzz:
 	mkdir -p ${corpusdir}
 	rm -rf ./go-fuzz-corpus
 	git clone https://github.com/dvyukov/go-fuzz-corpus.git ./go-fuzz-corpus/
+	go install golang.org/x/tools/cmd/file2fuzz@latest
 	file2fuzz -o ${corpusdir} ./go-fuzz-corpus/json/corpus/*  ./corpus/* 
 
 run:
-	GOARCH=amd64 go test -fuzz=${testname} -v
+	SONIC_FUZZ_MEM_LIMIT=2 GOMAXPROCS=4 GOARCH=amd64 go test -fuzz=${testname} -v -fuzztime 5m
 
 clean:
 	rm -rf ./go-fuzz-corpus/